modules/modeling/requirements.txt
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
torch==2.1.1
pytorch-lightning==2.1.2
pytorch-forecasting==1.0.0
joblib==1.3.2
optuna==3.5.0
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
